requirements.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
decorator==5.0.9
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
ipykernel==5.5.5
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2021.5.29
platformdirs==2.2.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pyzmq==22.1.0
six==1.16.0
sqlparse==0.4.1
tornado==6.1
traitlets==5.0.5
unicorn==1.0.3
virtualenv==20.7.2
virtualenv-clone==0.5.6
wcwidth==0.2.5
whitenoise==5.3.0
